miloco_server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "miloco-server"
version = "0.0.1"
description = "High-performance microservice system built on FastAPI, supporting smart home, AI chat, and MCP protocol features."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Miloco team", email = "xiaomi-miloco@xiaomi.com"}
]
dependencies = [
    "fastapi>=0.115.3",
    "uvicorn[standard]>=0.24.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.18",
    "pydantic>=2.4.0",
    "PyJWT>=2.8.0",
    "openai>=1.3.0",
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.3.0",
    "imagehash>=4.3.0",
    "fastmcp>=2.11",
    "aiohttp>=3.12.14",
    "thespian>=3.10.0",
    "aiofiles>=23.2.0",
    "aiocache>=0.12.0",
    "cachetools>=5.3.0",
    "croniter>=1.4.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "websockets>=12.0",
    "mem0ai>=0.1.0",
    "chromadb>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
miloco-server = "miloco_server.main:start_server"

[tool.setuptools.packages.find]
where = ["."]
include = ["miloco_server*"]

[tool.setuptools]
package-dir = {"" = "."}

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
